% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{as_integer_df}
\alias{as_integer_df}
\title{Convert all numeric columns in a dataframe to integers}
\usage{
as_integer_df(counts_tbl)
}
\arguments{
\item{counts_tbl}{data frame with numeric columns}
}
\value{
data frame with any numeric columns as integers
}
\description{
Round doubles to integers and convert to integer type
}
\examples{
\dontrun{
data.frame(a = c(0, 0.1, 2.3, 5L, 6.9)) \%>\% as_integer_df()
}
}
\keyword{internal}
