% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{calc_cpm}
\alias{calc_cpm}
\title{Calculate counts-per-million (CPM) on raw counts in a reneeDataSet}
\usage{
calc_cpm(renee_ds, ...)
}
\arguments{
\item{renee_ds}{reneeDataSet object}

\item{...}{additional arguments to pass to edgeR::cpm()}
}
\value{
reneeDataSet with cpm-transformed counts
}
\description{
Calculate counts-per-million (CPM) on raw counts in a reneeDataSet
}
\examples{
sample_meta <- data.frame(
  sample_id = c("KO_S3", "KO_S4", "WT_S1", "WT_S2"),
  condition = factor(
    c("knockout", "knockout", "wildtype", "wildtype"),
    levels = c("wildtype", "knockout")
  )
)
renee_ds <- create_reneeDataSet_from_dataframes(sample_meta, gene_counts) \%>\%
  calc_cpm()
head(renee_ds@counts$cpm)
}
