% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{calc_cpm_df}
\alias{calc_cpm_df}
\title{Calculate CPM on a data frame}
\usage{
calc_cpm_df(dat, gene_colname = "gene_id", ...)
}
\arguments{
\item{dat}{data frame of counts with a gene column}

\item{gene_colname}{name of the gene column (default: "gene_id")}

\item{...}{additional arguments to pass to edger::cpm()}
}
\value{
cpm-transformed counts as a data frame
}
\description{
Calculate CPM on a data frame
}
\keyword{internal}
