% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq2.R
\name{create_deseq_obj}
\alias{create_deseq_obj}
\title{Create DESeq2 object from gene counts and sample metadata}
\usage{
create_deseq_obj(counts_tbl, meta_dat, design)
}
\arguments{
\item{counts_tbl}{expected gene counts from RSEM as a data frame or tibble.}

\item{meta_dat}{sample metadata as a data frame with rownames as sample IDs.}

\item{design}{model formula for experimental design. Columns must exist in \code{meta_dat}.}
}
\value{
DESeqDataSet
}
\description{
Create DESeq2 object from gene counts and sample metadata
}
\examples{
sample_meta <- data.frame(
  row.names = c("KO_S3", "KO_S4", "WT_S1", "WT_S2"),
  condition = factor(c("knockout", "knockout", "wildtype", "wildtype"),
    levels = c("wildtype", "knockout")
  )
)
dds <- create_deseq_obj(gene_counts, sample_meta, ~condition)
}
