% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_renee-class.R
\name{create_reneeDataSet_from_dataframes}
\alias{create_reneeDataSet_from_dataframes}
\title{Construct a reneeDataSet object from data frames}
\usage{
create_reneeDataSet_from_dataframes(
  sample_meta_dat,
  count_dat,
  sample_id_colname = "sample_id",
  count_type = "raw"
)
}
\arguments{
\item{sample_meta_dat}{sample metadata as a data frame or tibble.
Must contain a \code{sample_ID} column.}

\item{count_dat}{expected gene counts from RSEM as a data frame or tibble.
Must contain a \code{gene_id} column and a column for each sample ID in the metadata.}
}
\value{
reneeDataSet object
}
\description{
Construct a reneeDataSet object from data frames
}
\examples{
sample_meta <- data.frame(
  sample_id = c("KO_S3", "KO_S4", "WT_S1", "WT_S2"),
  condition = factor(
    c("knockout", "knockout", "wildtype", "wildtype"),
    levels = c("wildtype", "knockout")
  )
)
create_reneeDataSet_from_dataframes(sample_meta, gene_counts)
}
