% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_counts}
\alias{filter_counts}
\title{Filter low counts}
\usage{
filter_counts(
  renee_ds,
  gene_names_column = "Gene",
  sample_names_column = "Sample",
  groups_column = "Group",
  labels_column = "Label",
  columns_to_include = c("Gene", "A1", "A2", "A3", "B1", "B2", "B3", "C1", "C2", "C3"),
  outlier_samples_to_remove = c(),
  use_cpm_counts_to_filter = TRUE,
  Minimum_Count_Value_to_be_Considered_Nonzero = 8,
  Minimum_Number_of_Samples_with_Nonzero_Counts_in_Total = 7,
  Use_Group_Based_Filtering = FALSE,
  Minimum_Number_of_Samples_with_Nonzero_Counts_in_a_Group = 3,
  principal_component_on_x_axis = 1,
  principal_component_on_y_axis = 2,
  legend_position_for_pca = "top",
  point_size_for_pca = 1,
  add_labels_to_pca = TRUE,
  label_font_size = 3,
  label_offset_y_ = 2,
  label_offset_x_ = 2,
  samples_to_rename_manually = c(""),
  color_histogram_by_group = FALSE,
  set_min_max_for_x_axis_for_histogram = FALSE,
  minimum_for_x_axis_for_histogram = -1,
  maximum_for_x_axis_for_histogram = 1,
  legend_position_for_histogram = "top",
  legend_font_size_for_histogram = 10,
  number_of_histogram_legend_columns = 6,
  colors_for_plots = c("indigo", "carrot", "lipstick", "turquoise", "lavender", "jade",
    "coral", "azure", "green", "rum", "orange", "olive"),
  number_of_image_rows = 2,
  interactive_plots = FALSE,
  plot_correlation_matrix_heatmap = TRUE
)
}
\arguments{
\item{counts_matrix}{The input Counts Matrix. Usually, this will be your Cleaned Counts matrix.}

\item{sample_metadata}{The Sample Metadata table containing your sample metadata. At minimum, this table must include one column each of the following: Samples, Groups, Batches, and Labels. The names in the Samples column of your input Sample Metadata must match the Sample Column Names of your input Counts Matrix exactly. You may have more than one column showing different Groups by which your samples may be organized (e.g. Genotype, Response, Time, etc.).}
}
\value{
Filtered counts as a dataframe
}
\description{
Filter low counts
}
\examples{
renee_ds <- create_reneeDataSet_from_dataframes(
  as.data.frame(nidap_sample_metadata),
  as.data.frame(nidap_clean_raw_counts),
  sample_id_colname = "Sample"
)
set.seed(10)
renee_ds2 <- filter_counts(renee_ds)
head(renee_ds2@counts[["filt"]])

}
