% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_counts}
\alias{gene_counts}
\title{RSEM gene counts}
\format{
\subsection{\code{gene_counts}}{

A data frame with columns 'gene_id', 'GeneName', and a column for each sample's count.
}
}
\source{
Generated by running RENEE v2.5.8 on the
\href{https://github.com/CCBR/RENEE/tree/e08f7db6c6e638cfd330caa182f64665d2ef37fa/.tests}{test dataset}
}
\usage{
gene_counts
}
\description{
RSEM gene counts
}
\keyword{datasets}
