% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{get_random_colors}
\alias{get_random_colors}
\title{Get random colors.}
\usage{
get_random_colors(num_colors, n = 2000)
}
\arguments{
\item{num_colors}{number of colors to select.}

\item{n}{number of random RGB values to generate in the color space.}
}
\value{
vector of random colors in hex format.
}
\description{
Note: this function is not guaranteed to create a color blind friendly palette.
Consider using other palettes such as \code{RColorBrewer::display.brewer.all(colorblindFriendly = TRUE)}.
}
\examples{
\dontrun{
set.seed(10)
get_random_colors(5)
}
}
\keyword{internal}
