% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nidap_clean_raw_counts}
\alias{nidap_clean_raw_counts}
\title{Clean raw counts for the NIDAP test dataset.
Pairs with \code{nidap_sample_metadata}.}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 43280 rows and 10 columns.
}
\usage{
nidap_clean_raw_counts
}
\description{
Clean raw counts for the NIDAP test dataset.
Pairs with \code{nidap_sample_metadata}.
}
\keyword{datasets}
