% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nidap_filtered_counts}
\alias{nidap_filtered_counts}
\title{Filtered counts for the NIDAP test dataset.
The result of running \code{filter_counts()} on \code{nidap_clean_raw_counts}.}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 7943 rows and 10 columns.
}
\usage{
nidap_filtered_counts
}
\description{
Filtered counts for the NIDAP test dataset.
The result of running \code{filter_counts()} on \code{nidap_clean_raw_counts}.
}
\keyword{datasets}
