% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nidap_sample_metadata}
\alias{nidap_sample_metadata}
\title{Sample metadata for the NIDAP test dataset}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 9 rows and 5 columns.
}
\usage{
nidap_sample_metadata
}
\description{
Sample metadata for the NIDAP test dataset
}
\keyword{datasets}
