% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{plot_pca}
\alias{plot_pca}
\title{Perform and plot a Principal Components Analysis}
\usage{
plot_pca(
  log_counts,
  sample_metadata,
  samples_to_include,
  samples_to_rename_manually,
  groups_column,
  labels_column,
  color_values,
  principal_component_on_x_axis = 1,
  principal_component_on_y_axis = 2,
  legend_position_for_pca = "top",
  point_size_for_pca = 1,
  add_labels_to_pca = TRUE,
  label_font_size = 3,
  label_offset_y_ = 2,
  label_offset_x_ = 2
)
}
\arguments{
\item{log_counts}{log-transformed filtered counts}

\item{sample_metadata}{The Sample Metadata table containing your sample metadata. At minimum, this table must include one column each of the following: Samples, Groups, Batches, and Labels. The names in the Samples column of your input Sample Metadata must match the Sample Column Names of your input Counts Matrix exactly. You may have more than one column showing different Groups by which your samples may be organized (e.g. Genotype, Response, Time, etc.).}

\item{samples_to_include}{samples in \code{sample_metadata} to include in the analysis}
}
\value{
PCA plot
}
\description{
Perform and plot a Principal Components Analysis
}
\keyword{internal}
