% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_raw_counts.R
\name{read_raw_counts}
\alias{read_raw_counts}
\title{read_raw_counts}
\usage{
read_raw_counts(pathtofile, delimiter = "\\t")
}
\arguments{
\item{pathtofile}{string absolute file path to raw counts matrix}

\item{delimiter}{string default is tab}
}
\value{
raw_counts_matrix
}
\description{
read_raw_counts
}
