% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{remove_low_count_genes}
\alias{remove_low_count_genes}
\title{Remove low-count genes}
\usage{
remove_low_count_genes(
  counts_matrix,
  sample_metadata,
  gene_names_column,
  groups_column,
  use_cpm_counts_to_filter = TRUE,
  Use_Group_Based_Filtering = FALSE,
  Minimum_Count_Value_to_be_Considered_Nonzero = 8,
  Minimum_Number_of_Samples_with_Nonzero_Counts_in_Total = 7,
  Minimum_Number_of_Samples_with_Nonzero_Counts_in_a_Group = 3
)
}
\arguments{
\item{counts_matrix}{The input Counts Matrix. Usually, this will be your Cleaned Counts matrix.}

\item{sample_metadata}{The Sample Metadata table containing your sample metadata. At minimum, this table must include one column each of the following: Samples, Groups, Batches, and Labels. The names in the Samples column of your input Sample Metadata must match the Sample Column Names of your input Counts Matrix exactly. You may have more than one column showing different Groups by which your samples may be organized (e.g. Genotype, Response, Time, etc.).}
}
\value{
counts matrix with low-count genes removed
}
\description{
TODO this function also transforms raw counts to CPM, but that should be a separate function before this step
TODO document \code{isexpr1} column in output
}
\keyword{internal}
