% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{rename_samples}
\alias{rename_samples}
\title{Rename samples}
\usage{
rename_samples(dat, samples_to_rename_manually)
}
\arguments{
\item{dat}{data frame containing a \code{sample} column}

\item{samples_to_rename_manually}{TODO use sample metadata spreadsheet custom column}
}
\value{
data frame with samples renamed
}
\description{
TODO this should happen right at the beginning of the template?
}
\details{
TODO accept new names for samples in sample metadata spreadsheet
}
\keyword{internal}
