% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renee-class.R
\name{reneeDataSetFromDataFrames}
\alias{reneeDataSetFromDataFrames}
\title{Construct a reneeDataSet object from data frames}
\usage{
reneeDataSetFromDataFrames(gene_counts_dat, sample_meta_dat)
}
\arguments{
\item{gene_counts_dat}{expected gene counts from RSEM as a data frame or tibble.
Must contain a \code{gene_id} column and a column for each sample ID in the metadata.}

\item{sample_meta_dat}{sample metadata as a data frame or tibble.
Must contain a \code{sample_ID} column.}
}
\value{
reneeDataSet object
}
\description{
Construct a reneeDataSet object from data frames
}
\examples{
sample_meta <- data.frame(
  sample_id = c("KO_S3", "KO_S4", "WT_S1", "WT_S2"),
  condition = factor(
    c("knockout", "knockout", "wildtype", "wildtype"),
    levels = c("wildtype", "knockout")
  )
)
reneeDataSetFromDataFrames(gene_counts, sample_meta)
}
