% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reneeTools-package.R
\docType{package}
\name{reneeTools-package}
\alias{reneeTools}
\alias{reneeTools-package}
\title{reneeTools: R helper functions for RENEE}
\description{
\code{reneeTools} implements helper functions for RENEE, a comprehensive
quality-control and quantification RNA-seq pipeline
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/CCBR/reneeTools}
  \item \url{https://ccbr.github.io/reneeTools/}
  \item Report bugs at \url{https://github.com/CCBR/reneeTools/issues}
}

}
\author{
\strong{Maintainer}: Vishal Koparde \email{vishal.koparde@nih.gov} (\href{https://orcid.org/0000-0001-8978-8495}{ORCID})

Authors:
\itemize{
  \item Kelly Sovacool \email{kelly.sovacool@nih.gov} (\href{https://orcid.org/0000-0003-3283-829X}{ORCID})
  \item Samantha Chill \email{samantha.chill@nih.gov} (\href{https://orcid.org/0000-0002-8734-9875}{ORCID})
}

Other contributors:
\itemize{
  \item CCR Collaborative Bioinformatics Resource [copyright holder]
}

}
\keyword{internal}
