% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_sample_metadata}
\alias{validate_sample_metadata}
\title{Validate raw counts sample metadata for filtering}
\usage{
validate_sample_metadata(
  counts_matrix,
  sample_metadata,
  sample_names_column = "Sample",
  groups_column = "Group"
)
}
\arguments{
\item{counts_matrix}{The input Counts Matrix. Usually, this will be your Cleaned Counts matrix.}

\item{sample_metadata}{The Sample Metadata table containing your sample metadata. At minimum, this table must include one column each of the following: Samples, Groups, Batches, and Labels. The names in the Samples column of your input Sample Metadata must match the Sample Column Names of your input Counts Matrix exactly. You may have more than one column showing different Groups by which your samples may be organized (e.g. Genotype, Response, Time, etc.).}
}
\value{
sample metadata with empty cells removed and special characters replaced
}
\description{
Validate raw counts sample metadata for filtering
}
\keyword{internal}
