% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renee-class.R
\name{create_reneeDataSet_from_files}
\alias{create_reneeDataSet_from_files}
\title{Construct a reneeDataSet object from tsv files.}
\usage{
create_reneeDataSet_from_files(gene_counts_filepath, sample_meta_filepath)
}
\arguments{
\item{gene_counts_filepath}{path to tsv file of expected gene counts from RSEM.}

\item{sample_meta_filepath}{path to tsv file with sample IDs and metadata for differential analysis.}
}
\value{
reneeDataSet object
}
\description{
Construct a reneeDataSet object from tsv files.
}
\examples{
create_reneeDataSet_from_files(
  system.file("extdata", "RSEM.genes.expected_count.all_samples.txt", package = "reneeTools"),
  system.file("extdata", "sample_metadata.tsv", package = "reneeTools")
)
}
