% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_low_counts.R
\name{filter_low_counts}
\alias{filter_low_counts}
\title{filter_low_counts}
\usage{
filter_low_counts(counts_dat, min_count = 0)
}
\arguments{
\item{counts_dat}{expected gene counts from RSEM as a data frame or tibble}

\item{min_count}{integer number of minimum counts across all samples (default: 0)}
}
\value{
filtered counts dataframe
}
\description{
filter_low_counts
}
\examples{
filter_low_counts(gene_counts) \%>\% head()
filter_low_counts(gene_counts, min_counts = 100)
}
