% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{meta_tbl_to_dat}
\alias{meta_tbl_to_dat}
\title{Convert sample metadata from a tibble to a dataframe with sample IDs as row names}
\usage{
meta_tbl_to_dat(meta_tbl)
}
\arguments{
\item{meta_tbl}{tibble with \code{sample_id} column}
}
\value{
dataframe where row names are the sample IDs
}
\description{
Convert sample metadata from a tibble to a dataframe with sample IDs as row names
}
\examples{
sample_meta_tbl <- readr::read_tsv(system.file("extdata",
  "sample_metadata.tsv",
  package = "reneeTools"
))
head(sample_meta_tbl)
meta_tbl_to_dat(sample_meta_tbl)
}
