% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renee-class.R
\name{reneeDataSetFromFiles}
\alias{reneeDataSetFromFiles}
\title{Create a reneeDataSet object from TSV files}
\usage{
reneeDataSetFromFiles(gene_counts_filepath, sample_meta_filepath)
}
\arguments{
\item{gene_counts_filepath}{path to TSV file of expected gene counts from RSEM}

\item{sample_meta_filepath}{path to TSV file with sample IDs and metadata for differential analysis}
}
\value{
reneeDataSet object
}
\description{
Create a reneeDataSet object from TSV files
}
\examples{
reneeDataSetFromFiles(
  system.file("extdata", "RSEM.genes.expected_count.all_samples.txt", package = "reneeTools"),
  system.file("extdata", "sample_metadata.tsv", package = "reneeTools")
)
}
