% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq2.R
\name{run_deseq2}
\alias{run_deseq2}
\title{Run DESeq2 on a reneeDataSet}
\usage{
run_deseq2(renee_ds, design, ...)
}
\arguments{
\item{renee_ds}{reneeDataSet object}

\item{design}{model formula for experimental design. Columns must exist in \code{meta_dat}.}

\item{...}{remaining variables are forwarded to \code{DESeq2::DESeq()}.}
}
\value{
reneeDataSet object with DESeq2 slot filled
}
\description{
Run DESeq2 on a reneeDataSet
}
\examples{
\dontrun{
renee_ds <- create_reneeDataSet_from_files(
  system.file("extdata", "sample_metadata.tsv",
    package = "reneeTools"
  ),
  system.file("extdata",
    "RSEM.genes.expected_count.all_samples.txt",
    package = "reneeTools"
  )
) \%>\% filter_counts()
renee_ds <- run_deseq2(renee_ds, ~condition)
}
}
