% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_sample_metadata}
\alias{validate_sample_metadata}
\title{Validate raw counts sample metadata for filtering}
\usage{
validate_sample_metadata(
  counts_matrix,
  sample_metadata,
  sample_names_column = "Sample",
  group_column = "Group"
)
}
\arguments{
\item{sample_names_column}{The column from your input Sample Metadata table containing the sample names. The names in this column must exactly match the names used as the sample column names of your input Counts Matrix. Only columns of Text type from your input Sample Metadata table will be available to select for this parameter.}

\item{group_column}{The column from your input Sample Metadata table containing the sample group information. This is usually a column showing to which experimental treatments each sample belongs (e.g. WildType, Knockout, Tumor, Normal, Before, After, etc.). Only columns of Text type from your input Sample Metadata will be available to select for this parameter.}
}
\value{
sample metadata with empty cells removed and special characters replaced
}
\description{
Validate raw counts sample metadata for filtering
}
\keyword{internal}
